moved analytics to separate route
