+++ client/src/components/StateRatio.js
@@ -48,7 +48,7 @@ export default class StateRatio extends Component {
 
   getRatio = async values => {
     var startTime = new Date();
-    const result = await axios.post("http://localhost:9000/datasetpoint/getratio", values);
+    const result = await axios.post("http://localhost:9000/analytics/getratio", values);
     console.log(typeof result.data);
     this.setState({ratio: result.data});
     var endTime = new Date();
