+++ client/src/components/byYear.js
@@ -47,7 +47,7 @@ export default class MostPopular2 extends Component {
   };
 
   getbyYear = async values => {
-    const result = await axios.post("http://localhost:9000/datasetpoint/byYear", values);
+    const result = await axios.post("http://localhost:9000/analytics/byYear", values);
     console.log(result.data);
     this.setState({launched: result.data});
     this.setState({total: result.data.total});
